pyproject.toml
[tool.poetry]
name = "zephir-api"
version = "0.1.0"
description = "Zephir Item API"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0"
flask_sqlalchemy = "^3.0"
mysql-connector-python = "^8.4.0"
markdown = "^3.6"
markupsafe = "^2.1.5"
gunicorn = "^22.0.0"
boto3 = "^1.34.154"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-flask = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"